* Select the "All Fonts" group by default - event in non-management mode * Don't show an error dialog with an empty string. * Use quint32 to store fontconfig style information.
svn path=/trunk/KDE/kdebase/workspace/; revision=700448

+++ kcontrol/kfontinst/kio/KioFonts.h
@@ -88,7 +88,7 @@ class CKioFonts : public KIO::SlaveBase
         TFontDetails() : writingSystems(0) { }
 
         CDisabledFonts::TFileList files;
-        unsigned long             styleVal;
+        quint32                   styleVal;
         qulonglong                writingSystems;
     };
 
@@ -140,7 +140,7 @@ class CKioFonts : public KIO::SlaveBase
     QString            getGroupName(gid_t gid);
     bool               createFontUDSEntry(KIO::UDSEntry &entry, const QString &name,
                                           const CDisabledFonts::TFileList &patterns,
-                                          unsigned long styleVal, qulonglong writingSystems,
+                                          quint32 styleVal, qulonglong writingSystems,
                                           bool sys, bool hidden=false);
     bool               createFolderUDSEntry(KIO::UDSEntry &entry, const QString &name, const QString &path,
                                             bool sys);
